+++ src/cmake/AxomOptions.cmake
@@ -56,9 +56,9 @@ option(AXOM_ENABLE_TOOLS "Enables Axom Tools" ON)
 option(AXOM_ENABLE_TUTORIALS "Builds Axom tutorials as part of the Axom build" ON)
 mark_as_advanced(AXOM_ENABLE_TUTORIALS)
 
-option(AXOM_ENABLE_SLIC_DEBUG_MACROS
-       "Enables SLIC_ASSERT, SLIC_CHECK, and SLIC_DEBUG macros independently of AXOM_DEBUG"
-       OFF)
+set(AXOM_ENABLE_SLIC_DEBUG_MACROS "DEFAULT" CACHE STRING
+    "Controls SLIC_ASSERT, SLIC_CHECK, and SLIC_DEBUG macros independently of AXOM_DEBUG")
+set_property(CACHE AXOM_ENABLE_SLIC_DEBUG_MACROS PROPERTY STRINGS "DEFAULT" "ON" "OFF")
 
 #------------------------------------------------------------------------------
 # Test execution controls